+++ python/CHANGELOG.md
@@ -8,6 +8,7 @@ We structure this changelog in accordance with [Keep a Changelog](https://keepac
 
 ### Changed
 - **BREAKING**: `BatchLoader.get_batch()` no longer takes `extractor` and `decoder` args. Instead, use `return_raw` and `decode_as_stream`.
+- Support `ETLConfig.args` parameter in `Object.copy` method.
 
 ## [1.15.0] - 2025-07-15